edit 100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -8,7 +8,7 @@
  */
 int _atoi(char *s)
 {
-	long int num = 0, n = 0;
+	double num = 0, n = 0;
 
 	while (*s != '\0')
 	{
